Formateo FULL y FAST creado correctamente. Sistema LWH instalado correctamente. Se personalizó la terminal. Algunos ajustes en estructs.

+++ Proyecto1/estructuras/dd.go
@@ -3,13 +3,13 @@ package estructuras
 //DD struct
 type DD struct {
 	DDFiles     [5]DDFile
-	ApuntadorDD int64
+	ApuntadorDD int32
 }
 
 //DDFile struct
 type DDFile struct {
 	Name              [20]byte
-	ApuntadorInodo    int64
+	ApuntadorInodo    int32
 	FechaCreacion     [20]byte
 	FechaModificacion [20]byte
 }
